refactor: replace GitHubDiffProvider interface with GitHubProvider interface
